+++ src/App.css
@@ -7,6 +7,10 @@
 
 .card {
   padding: 2em;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 .weather {
@@ -19,33 +23,31 @@
   padding: 20px;
   border-radius: 10px;
   color: white;
-  margin: 0 auto;
   margin-bottom: 20px;
 }
 
 .weather-card {
-  max-width: 300px;
+  max-width: 240px;
   width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
   background-color: #2467e4;
-  padding: 20px;
+  padding: 10px;
   border-radius: 10px;
   color: white;
-  margin-bottom: 20px;
 }
 
 .info {
   display: flex;
   align-items: center;
   flex-direction: column;
-  font-size: 14px;
+  font-size: 12px;
 }
 
 .info span {
   font-weight: 700;
-  font-size: 40px;
+  font-size: 30px;
   display: block;
 }
 
@@ -62,8 +64,13 @@
 
 .date {
   margin-bottom: 20px;
+  font-size: 24px;
 }
 
 .form-title {
   font-size: 1.8rem;
+}
+
+.search {
+  margin-bottom: 40px;
 }
