+++ src/index.css
@@ -1,3 +1,9 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -14,6 +20,7 @@
 }
 
 
+
 body {
   margin: 0;
   display: flex;
@@ -25,6 +32,7 @@ body {
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
+  margin-bottom: 20px;
 }
 
 p {
@@ -53,4 +61,8 @@ button {
   font-size: 1.2em;
   cursor: pointer;
   margin-top: 20px;
+}
+
+button:hover {
+  background-color: #2a41e8;
 }
